Rename withLoading HOC to withExecutionLoading
